Rename and split ValidateConfig files
